\name{dataInteratrix}
\alias{dataInteratrix}

\docType{data}
\title{A generated data set for test}

\description{A generated data set provided to test the \code{Interatrix} package.}

\usage{data(dataInteratrix)}

\format{
  A data frame with 100 observations for the following variables:
  \describe{
    \item{\code{F1}}{a numeric vector containing a factor with three modalities}
    \item{\code{F2}}{a numeric vector containing a continuous variable}
    \item{\code{F3}}{a numeric vector containing a factor with two modalities}
    \item{\code{F4}}{a numeric vector containing a continuous variable}
    \item{\code{Parasite1}}{a numeric vector containing the serological status to the first parasite}
    \item{\code{Parasite2}}{a numeric vector containing the serological status to the second parasite}
    \item{\code{AGE}}{a numeric vector containing a factor with three modalities indicating the age classes}
  }
}

\examples{data(dataInteratrix)}

\keyword{datasets}
