\name{InteratrixGUI}
\alias{InteratrixGUI}

\title{Function to start the graphical interface}
\description{This function opens a graphical interface and helps step by step to compute corrected chi-square tests.}

\usage{InteratrixGUI()}

\value{A first interactive graphical interface is opened to choose between two methods.
When all parameters are defined by the user, simulation results are printed to the 
R console, saved in a file and plotted as an histogram.}


\keyword{hplot}
\keyword{print}
\keyword{htest}
\keyword{interface}

