\name{Interatrix-package}
\alias{Interatrix-package}
\alias{Interatrix}

\docType{package}
\title{Compute chi-square tests with corrections}

\description{Compute chi-square tests with corrections}

\details{
\tabular{ll}{
Package: \tab Interatrix\cr
Type: \tab Package\cr
Version: \tab 1.1.1\cr
Date: \tab 2015-07-03\cr
License: \tab GPL (>= 2)\cr
}
}
\author{Authors: Aurélie Siberchicot<aurelie.siberchicot@univ-lyon1.fr>, Eléonore Hellard, Dominique Pontier, David Fouchet and Franck Sauvage}

\keyword{package}
